This Works awsome

+++ Order_API/DTO/OrderDTO.cs
@@ -49,10 +49,11 @@ namespace Order_API.DTO
         public int Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public int MenuCategoryId { get; set; }
         public string Image { get; set; }
     }
+
     public class ExtraItemDto
     {
         public int Id { get; set; }
